# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Live API
  version: 1.0.0
  description: |
    REST API for the Open Live broadcast production platform.

    **Authentication**: Pass an API key via `Authorization: Bearer <key>` header or `?key=<key>` query parameter.
    The following paths are exempt from authentication: `/health`, `/ready`, `/api/v1/status`, `/api/v1/server-info`, `/api/v1/reconnect`.
servers:
  - url: /
security:
  - bearerAuth: []
  - apiKeyQuery: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyQuery:
      type: apiKey
      in: query
      name: key
  schemas:
...
    StreamType:
      type: string
      enum: [srt, efp, whip, html]
    SourceStatus:
      type: string
      enum: [active, inactive]
    OutputType:
      type: string
      enum: [mpegtssrt, efpsrt, whep]
...
    Source:
      type: object
      required: [id, name, address, streamType, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000
          description: SRT latency in milliseconds
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SourceInput:
      type: object
      required: [name, address, streamType]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
          default: inactive
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000
    SourcePatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000

    # ── Graphic ─────────────────────────────────────────────────────────────────
...
    Output:
      type: object
      required: [id, name, outputType, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        outputType:
          $ref: '#/components/schemas/OutputType'
        url:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OutputInput:
      type: object
      required: [name, outputType]
      properties:
        name:
          type: string
          minLength: 1
        outputType:
          $ref: '#/components/schemas/OutputType'
        url:
          type: string
    OutputPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        url:
          type: string

    # ── Production Config ────────────────────────────────────────────────────────
...
    Pipeline:
      type: object
      required: [status]
      properties:
        stromConfig:
          type: object
          nullable: true
          additionalProperties: true
        status:
          type: string
          enum: [stopped, running]
...
    Error:
      type: object
      properties:
        error:
          type: string
        statusCode:
          type: integer
paths:
...
  /api/v1/sources:
    get:
      summary: List sources
      operationId: listSources
      tags: [Sources]
      responses:
        '200':
          description: Array of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Source'
    post:
      summary: Create source
      operationId: createSource
      tags: [Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceInput'
      responses:
        '201':
          description: Created source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/sources/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get source
      operationId: getSource
      tags: [Sources]
      responses:
        '200':
          description: Source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Not found
    patch:
      summary: Update source
      operationId: updateSource
      tags: [Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcePatch'
      responses:
        '200':
          description: Updated source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Not found
    delete:
      summary: Delete source
      operationId: deleteSource
      tags: [Sources]
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found

  # ── Graphics ─────────────────────────────────────────────────────────────────
...
  /api/v1/outputs:
    get:
      summary: List outputs
      operationId: listOutputs
      tags: [Outputs]
      responses:
        '200':
          description: Array of outputs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Output'
    post:
      summary: Create output
      operationId: createOutput
      tags: [Outputs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutputInput'
      responses:
        '201':
          description: Created output
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Output'
        '400':
          description: Validation error
  /api/v1/outputs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get output
      operationId: getOutput
      tags: [Outputs]
      responses:
        '200':
          description: Output
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Output'
        '404':
          description: Not found
    patch:
      summary: Update output
      operationId: updateOutput
      tags: [Outputs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutputPatch'
      responses:
        '200':
          description: Updated output
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Output'
        '404':
          description: Not found
    delete:
      summary: Delete output
      operationId: deleteOutput
      tags: [Outputs]
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found

  # ── Production Configs ───────────────────────────────────────────────────────
...
  /api/v1/productions/{id}/pipeline:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get pipeline state
      operationId: getPipeline
      tags: [Pipeline]
      responses:
        '200':
          description: Pipeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
        '404':
          description: Not found
  /api/v1/productions/{id}/pipeline/config:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      summary: Update pipeline Strom config
      operationId: updatePipelineConfig
      tags: [Pipeline]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stromConfig]
              properties:
                stromConfig:
                  type: object
                  additionalProperties:
                    oneOf:
                      - type: string
                        maxLength: 1024
                      - type: number
                      - type: boolean
                      - type: 'null'
      responses:
        '200':
          description: Updated pipeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
  /api/v1/productions/{id}/pipeline/start:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Start pipeline
      operationId: startPipeline
      tags: [Pipeline]
      responses:
        '200':
          description: Pipeline started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
  /api/v1/productions/{id}/pipeline/stop:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Stop pipeline
      operationId: stopPipeline
      tags: [Pipeline]
      responses:
        '200':
          description: Pipeline stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'

  # Production — Audio
...